Did some cleanup
