+++ src/lib.rs
@@ -1,3 +1,2 @@
 pub mod types;
-pub mod plugin_info;
-pub mod initialized_plugin;
+pub mod plugin;
