+++ src/plugin/api_version.rs
@@ -1,5 +1,5 @@
 #[repr(C)]
-#[derive(Copy, Clone)]
+#[derive(Copy, Clone, Debug)]
 pub struct APIVersion {
     pub major: i32,
     pub minor: i32,
